Refactoring: moved tests for representative licenses into new module
